chore(console): rename old jest setup for angularJs disabled test
