+++ gravitee-apim-console-webui/src/components/alerts/alert/triggers/conditions/trigger-condition-string.component.ajs-spec.ts
@@ -15,7 +15,7 @@
  */
 import { IComponentControllerService } from 'angular';
 
-import { setupAngularJsTesting } from '../../../../../../jest.setup';
+import { setupAngularJsTesting } from '../../../../../../old-jest.setup';
 import { Metrics, Scope } from '../../../../../entities/alert';
 
 setupAngularJsTesting();
