+++ gravitee-apim-console-webui/src/management/api/analytics/logs/analytics-log.component.ajs-spec.ts
@@ -15,7 +15,7 @@
  */
 import { IComponentControllerService } from 'angular';
 
-import { setupAngularJsTesting } from '../../../../../jest.setup.js';
+import { setupAngularJsTesting } from '../../../../../old-jest.setup.js';
 
 setupAngularJsTesting();
