+++ gravitee-apim-console-webui/src/services/event.service.ajs-spec.ts
@@ -18,7 +18,7 @@ import { IHttpBackendService } from 'angular';
 import { EventService } from './event.service';
 
 import { fakeEvent } from '../entities/event/event.fixture';
-import { setupAngularJsTesting } from '../../jest.setup.js';
+import { setupAngularJsTesting } from '../../old-jest.setup.js';
 
 setupAngularJsTesting();
