+++ gravitee-apim-console-webui/src/services/qualityRule.service.ajs-spec.ts
@@ -19,7 +19,7 @@ import { IHttpBackendService } from 'angular';
 import QualityRuleService from './qualityRule.service';
 
 import { QualityRule } from '../entities/qualityRule';
-import { setupAngularJsTesting } from '../../jest.setup.js';
+import { setupAngularJsTesting } from '../../old-jest.setup.js';
 
 setupAngularJsTesting();
